+++ src/components/features/product/index.ts
@@ -1,0 +1,3 @@
+export { default as ProductCard } from './product-card/ProductCard';
+export { default as ProductRowCard } from './product-card/ProductRowCard';
+export type * from './type';
