+++ src/pages/shop/context/index.ts
@@ -1,0 +1,3 @@
+export { default as ShopErrorProvider } from './ShopErrorProvider';
+export { ShopErrorContext } from './ShopErrorProvider';
+export { useShopErrorContext } from './useShopErrorContext';
